+++ src/server/matchers/sources.rs
@@ -311,32 +311,6 @@ impl MultiValueSource<String, String> for ContainsQueryParameterSource {
     }
 }
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 // ************************************************************************************************
 // QueryParameterSource
 // ************************************************************************************************
@@ -380,35 +354,6 @@ impl MultiValueSource<String, String> for ContainsXWWWFormUrlencodedKeySource {
             .map(|v| v.into_iter().map(|v| (v, None)).collect())
     }
 }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 
 // ************************************************************************************************
 // FunctionSource