+++ tests/examples/mod.rs
@@ -13,4 +13,4 @@ mod showcase_tests;
 mod standalone_tests;
 mod string_body_tests;
 mod url_matching_tests;
-mod x_www_form_urlencoded_body_tests;
+mod x_www_form_urlencoded_tests;